Added review endpoint with Dtos and simple CRUD operations

+++ src/main/java/com/app/oslotoilet/toilet/ToiletService.java
@@ -68,6 +68,9 @@ public class ToiletService {
 
 
     public void deleteToilet(UUID id){
+        if (!toiletRepository.existsById(id)){
+            throw new EntityNotFoundException("Toilet not found");
+        }
         toiletRepository.deleteById(id);
     }
